+++ OpenGL/OBJ.cpp
@@ -17,7 +17,7 @@ OBJ::OBJ(float x, float y, float z)
 	_position.y = y;
 	_position.z = z;
 
-	_rotation = rand() % 360;
+	_rotation = (GLfloat)(rand() % 360);
 
 	_rotationSpeed = ((rand() % 19) + 2) / 10.0f;
 
@@ -178,7 +178,7 @@ void OBJ::Update()
 		if (_axisX == 0 && _axisY == 0 && _axisZ == 0)
 			_axisY = 1.0f;
 
-		_rotation = rand() % 360;
+		_rotation = (GLfloat)(rand() % 360);
 	}
 }
